collapsed ViolationIntegrator and CitationIntegrator into CaseIntegrator
