+++ src/main/java/com/tcvcog/tcvce/entities/CodeViolation.java
@@ -36,6 +36,7 @@ public  class       CodeViolation
     protected int ceCaseID;
        
     protected ViolationStatusEnum status;
+    protected boolean active;
     protected Icon icon;
     protected String ageLeadText;
     
@@ -58,25 +59,26 @@ public  class       CodeViolation
     
     protected LocalDateTime creationTS;
     protected String creationTSPretty;
-    
     protected User createdBy;
     
     protected LocalDateTime stipulatedComplianceDate;
-    
     protected LocalDateTime actualComplianceDate;
     
     protected boolean leagacyImport;
-    
     protected List<Integer> blobIDList;
+    protected List<Integer> photoList;
     
     protected LocalDateTime complianceTimeStamp;
     protected User complianceUser;
-    protected EventCnF compTimeFrameComplianceEvent;
-    protected int complianceTimeframeEventID;
     
-    protected List<Integer> photoList;
+    protected int complianceTFExpiryPropID;
+    protected Proposal complianceTFExpiryProp;
     
-    private int severityIntensityClassID;
+    protected int severityIntensityClassID;
+    
+    protected LocalDateTime lastUpdatedTS;
+    protected User lastUpdatedUser;
+    
     
     
      /**
@@ -397,12 +399,7 @@ public  class       CodeViolation
         return complianceUser;
     }
 
-    /**
-     * @return the compTimeFrameComplianceEvent
-     */
-    public EventCnF getCompTimeFrameComplianceEvent() {
-        return compTimeFrameComplianceEvent;
-    }
+   
 
     /**
      * @param leagacyImport the leagacyImport to set
@@ -425,26 +422,7 @@ public  class       CodeViolation
         this.complianceUser = complianceUser;
     }
 
-    /**
-     * @param compTimeFrameComplianceEvent the compTimeFrameComplianceEvent to set
-     */
-    public void setCompTimeFrameComplianceEvent(EventCnF compTimeFrameComplianceEvent) {
-        this.compTimeFrameComplianceEvent = compTimeFrameComplianceEvent;
-    }
-
-    /**
-     * @return the complianceTimeframeEventID
-     */
-    public int getComplianceTimeframeEventID() {
-        return complianceTimeframeEventID;
-    }
-
-    /**
-     * @param complianceTimeframeEventID the complianceTimeframeEventID to set
-     */
-    public void setComplianceTimeframeEventID(int complianceTimeframeEventID) {
-        this.complianceTimeframeEventID = complianceTimeframeEventID;
-    }
+    
     
     /**
      * Violations can print themselves on a single line
@@ -486,8 +464,6 @@ public  class       CodeViolation
         hash = 53 * hash + (this.leagacyImport ? 1 : 0);
         hash = 53 * hash + Objects.hashCode(this.complianceTimeStamp);
         hash = 53 * hash + Objects.hashCode(this.complianceUser);
-        hash = 53 * hash + Objects.hashCode(this.compTimeFrameComplianceEvent);
-        hash = 53 * hash + this.complianceTimeframeEventID;
         hash = 53 * hash + Objects.hashCode(this.blobIDList);
         return hash;
     }
@@ -664,6 +640,90 @@ public  class       CodeViolation
      */
     public void setSeverityIntensityClassID(int severityIntensityClassID) {
         this.severityIntensityClassID = severityIntensityClassID;
+    }
+
+    /**
+     * @return the complianceTFExpiryPropID
+     */
+    public int getComplianceTFExpiryPropID() {
+        return complianceTFExpiryPropID;
+    }
+
+    /**
+     * @return the complianceTFExpiryProp
+     */
+    public Proposal getComplianceTFExpiryProp() {
+        return complianceTFExpiryProp;
+    }
+
+    /**
+     * @return the photoList
+     */
+    public List<Integer> getPhotoList() {
+        return photoList;
+    }
+
+    /**
+     * @param complianceTFExpiryPropID the complianceTFExpiryPropID to set
+     */
+    public void setComplianceTFExpiryPropID(int complianceTFExpiryPropID) {
+        this.complianceTFExpiryPropID = complianceTFExpiryPropID;
+    }
+
+    /**
+     * @param complianceTFExpiryProp the complianceTFExpiryProp to set
+     */
+    public void setComplianceTFExpiryProp(Proposal complianceTFExpiryProp) {
+        this.complianceTFExpiryProp = complianceTFExpiryProp;
+    }
+
+    /**
+     * @param photoList the photoList to set
+     */
+    public void setPhotoList(List<Integer> photoList) {
+        this.photoList = photoList;
+    }
+
+    /**
+     * @return the lastUpdatedTS
+     */
+    public LocalDateTime getLastUpdatedTS() {
+        return lastUpdatedTS;
+    }
+
+    /**
+     * @return the lastUpdatedUser
+     */
+    public User getLastUpdatedUser() {
+        return lastUpdatedUser;
+    }
+
+    /**
+     * @param lastUpdatedTS the lastUpdatedTS to set
+     */
+    public void setLastUpdatedTS(LocalDateTime lastUpdatedTS) {
+        this.lastUpdatedTS = lastUpdatedTS;
+    }
+
+    /**
+     * @param lastUpdatedUser the lastUpdatedUser to set
+     */
+    public void setLastUpdatedUser(User lastUpdatedUser) {
+        this.lastUpdatedUser = lastUpdatedUser;
+    }
+
+    /**
+     * @return the active
+     */
+    public boolean isActive() {
+        return active;
+    }
+
+    /**
+     * @param active the active to set
+     */
+    public void setActive(boolean active) {
+        this.active = active;
     }
